sigstore: flatten models into sigstore.models
