+++ sigstore/_internal/merkle.py
@@ -33,7 +33,7 @@ from sigstore._utils import HexStr
 from sigstore.errors import VerificationError
 
 if typing.TYPE_CHECKING:
-    from sigstore.transparency import LogEntry
+    from sigstore.models import LogEntry
 
 
 _LEAF_HASH_PREFIX = 0
