+++ sigstore/sign.py
@@ -64,8 +64,8 @@ from sigstore._internal.rekor.client import RekorClient
 from sigstore._internal.sct import verify_sct
 from sigstore._internal.trustroot import KeyringPurpose, TrustedRoot
 from sigstore._utils import sha256_digest
+from sigstore.models import Bundle
 from sigstore.oidc import ExpiredIdentity, IdentityToken
-from sigstore.verify.models import Bundle
 
 _logger = logging.getLogger(__name__)
