+++ test/unit/test_models.py
@@ -12,12 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json
 from base64 import b64encode
 
 import pytest
 from pydantic import ValidationError
 
-from sigstore.transparency import LogEntry, LogInclusionProof
+from sigstore.models import Bundle, InvalidBundle, LogEntry, LogInclusionProof
 
 
 class TestLogEntry:
@@ -77,3 +78,40 @@ class TestLogInclusionProof:
                     tree_size=100,
                 ),
             )
+
+
+class TestBundle:
+    """
+    Tests for the `Bundle` wrapper model.
+    """
+
+    def test_invalid_bundle_version(self, signing_bundle):
+        with pytest.raises(InvalidBundle, match="unsupported bundle format"):
+            signing_bundle("bundle_invalid_version.txt")
+
+    def test_invalid_empty_cert_chain(self, signing_bundle):
+        with pytest.raises(
+            InvalidBundle, match="expected non-empty certificate chain in bundle"
+        ):
+            signing_bundle("bundle_no_cert_v1.txt")
+
+    def test_invalid_no_log_entry(self, signing_bundle):
+        with pytest.raises(
+            InvalidBundle, match="expected exactly one log entry in bundle"
+        ):
+            signing_bundle("bundle_no_log_entry.txt")
+
+    def test_verification_materials_offline_no_checkpoint(self, signing_bundle):
+        with pytest.raises(
+            InvalidBundle, match="expected checkpoint in inclusion proof"
+        ):
+            signing_bundle("bundle_no_checkpoint.txt")
+
+    def test_bundle_roundtrip(self, signing_bundle):
+        _, bundle = signing_bundle("bundle.txt")
+
+        # Bundles are not directly comparable, but a round-trip preserves their
+        # underlying object structure.
+        assert json.loads(Bundle.from_json(bundle.to_json()).to_json()) == json.loads(
+            bundle.to_json()
+        )
